+++ compliance/locker.py
@@ -16,12 +16,12 @@
 
 import json
 import logging
-import os
 import re
 import shutil
 import tempfile
 import time
 from datetime import datetime as dt, timedelta
+from pathlib import Path, PurePath
 from threading import Lock
 from urllib.parse import urlparse
 
@@ -44,6 +44,10 @@ DAY = 60 * 60 * 24
 AE_DEFAULT = 30 * DAY
 READMES = ['README.md', 'readme.md', 'Readme.md']
 AE_EXEMPT = [INDEX_FILE] + READMES
+NOT_EVIDENCE = AE_EXEMPT
+KB = 1000
+MB = KB * 1000
+LF_DEFAULT = 50 * MB
 
 
 class Locker(object):
@@ -107,10 +111,10 @@ class Locker(object):
         if name is not None:
             self.name = name
         elif repo_url is not None and name is None:
-            self.name = os.path.split(self.repo_url)[1]
+            self.name = repo_url.rsplit('/', 1)[1]
         elif repo_url is None and name is None:
             self.name = 'example'
-        self.local_path = os.path.join(tempfile.gettempdir(), self.name)
+        self.local_path = str(PurePath(tempfile.gettempdir(), self.name))
         self._do_push = do_push
         self.ttl_tolerance = ttl_tolerance
         self.gitconfig = gitconfig or {}
@@ -143,15 +147,14 @@ class Locker(object):
         ]
         touched_files = []
         for f in index_files:
-            index_dir = os.path.dirname(f)
-            content = json.loads(open(os.path.join(self.local_path, f)).read())
+            content = json.loads(Path(self.local_path, f).read_text())
             for t, desc in content.items():
                 if type(desc) is dict:
                     ts = desc.get('last_update')
                 else:
                     ts = desc
                 if ts == self.commit_date:
-                    touched_files.append(os.path.join(index_dir, t))
+                    touched_files.append(str(PurePath(f).with_name(t)))
         return touched_files
 
     def reset_depenency_rerun(self):
@@ -211,22 +214,16 @@ class Locker(object):
         if evidence.content is None:
             raise ValueError(f'Evidence {evidence.name} has no content')
 
-        if not os.path.isdir(os.path.join(self.local_path, evidence.dir_path)):
-            os.makedirs(os.path.join(self.local_path, evidence.dir_path))
+        path = Path(self.local_path, evidence.dir_path)
+        path.mkdir(parents=True, exist_ok=True)
         if getattr(evidence, 'is_partitioned', False):
             for key in evidence.partition_keys:
-                full_path = os.path.join(
-                    self.local_path,
-                    evidence.dir_path,
-                    f'{get_sha256_hash(key, 10)}_{evidence.name}'
-                )
-                with open(full_path, 'w') as f:
-                    f.write(evidence.get_partition(key))
+                ev_name = f'{get_sha256_hash(key, 10)}_{evidence.name}'
+                Path(path, ev_name).write_text(evidence.get_partition(key))
+        elif not getattr(evidence, 'binary_content', False):
+            Path(path, evidence.name).write_text(evidence.content)
         else:
-            full_path = os.path.join(self.local_path, evidence.path)
-            mode = 'wb' if getattr(evidence, 'binary_content', False) else 'w'
-            with open(full_path, mode) as f:
-                f.write(evidence.content)
+            Path(path, evidence.name).write_bytes(evidence.content)
         if not evidence.path.startswith('tmp/'):
             self.index(evidence, checks, evidence_used)
 
@@ -241,12 +238,11 @@ class Locker(object):
           applicable for check generated ReportEvidence.
         """
         with self.lock:
-            index_file = self.get_index_file(evidence)
-            repo_files = [index_file]
-            if not os.path.exists(index_file):
-                metadata = {}
-            else:
-                metadata = json.loads(open(index_file).read())
+            index_file = Path(self.get_index_file(evidence))
+            repo_files = [str(index_file)]
+            metadata = {}
+            if index_file.is_file():
+                metadata = json.loads(index_file.read_text())
             ev_meta = metadata.get(evidence.name, {})
             old_parts = ev_meta.get('partitions', {}).keys()
             metadata[evidence.name] = {
@@ -254,10 +250,12 @@ class Locker(object):
                 'ttl': evidence.ttl,
                 'description': evidence.description
             }
+            if evidence.is_empty:
+                metadata[evidence.name]['empty'] = True
             tombstones = None
             if getattr(evidence, 'is_partitioned', False):
                 unpartitioned = self.get_file(evidence.path)
-                if os.path.isfile(unpartitioned):
+                if Path(unpartitioned).is_file():
                     # Remove/tombstone unpartitioned evidence file
                     # replaced by partitioned evidence files
                     self.repo.index.remove([unpartitioned], working_tree=True)
@@ -309,8 +307,7 @@ class Locker(object):
                 metadata[evidence.name]['checks'] = checks
             if evidence_used is not None:
                 metadata[evidence.name]['evidence'] = evidence_used
-            with open(index_file, 'w') as f:
-                f.write(format_json(metadata))
+            index_file.write_text(format_json(metadata))
             self.repo.index.add(repo_files)
 
     def remove_partitions(self, evidence, hashes):
@@ -326,16 +323,9 @@ class Locker(object):
         """
         if not hashes:
             return
-        self.repo.index.remove(
-            [
-                os.path.join(
-                    self.local_path,
-                    evidence.dir_path,
-                    f'{part_hash}_{evidence.name}'
-                ) for part_hash in hashes
-            ],
-            working_tree=True
-        )
+        path = PurePath(self.local_path, evidence.dir_path)
+        ev_files = [str(path.joinpath(f'{h}_{evidence.name}')) for h in hashes]
+        self.repo.index.remove(ev_files, working_tree=True)
 
     def create_tombstone_metadata(self, candidate, metadata, reason):
         """
@@ -403,8 +393,8 @@ class Locker(object):
 
         :returns: the full path to the evidence's index file.
         """
-        path = evidence_path.rsplit('/', 1).pop(0)
-        return os.path.join(self.local_path, path, INDEX_FILE)
+        path = PurePath(self.local_path, evidence_path).with_name(INDEX_FILE)
+        return str(path)
 
     def get_file(self, filename):
         """
@@ -416,7 +406,7 @@ class Locker(object):
 
         :returns: the path to the filename provided.
         """
-        return os.path.join(self.local_path, filename)
+        return str(PurePath(self.local_path, filename))
 
     def get_remote_location(
         self, filename, include_commit=True, sha=None, locker_url=None
@@ -434,8 +424,7 @@ class Locker(object):
         :returns: the remote repository path to the filename provided.
         """
         if not self.repo_url_with_creds:
-            return os.path.join(self.local_path, filename)
-
+            return self.get_file(filename)
         ref = self.branch
         if include_commit:
             ref = self.repo.head.commit.hexsha
@@ -546,8 +535,7 @@ class Locker(object):
         :returns: True/False time to live validation.
         """
         if isinstance(evidence, TmpEvidence):
-            full_path = os.path.join(self.local_path, evidence.path)
-            return os.path.exists(full_path)
+            return Path(self.local_path, evidence.path).is_file()
         try:
             self._validate_evidence(evidence, ignore_ttl)
             return True
@@ -568,7 +556,7 @@ class Locker(object):
 
         :param locker: the locker "name" used in logging
         """
-        if os.path.isdir(os.path.join(self.local_path, '.git')):
+        if Path(self.local_path, '.git').is_dir():
             self.logger.info(f'Using {locker} found in {self.local_path}...')
             if not hasattr(self, 'repo'):
                 self.repo = git.Repo(self.local_path)
@@ -632,6 +620,7 @@ class Locker(object):
             )
             remote.fetch()
             remote.pull(rebase=True)
+            self._log_large_files()
             self.logger.info(
                 f'Pushing local locker to remote repo {self.repo_url}...'
             )
@@ -641,36 +630,33 @@ class Locker(object):
 
     def add_content_to_locker(self, content, folder='', filename=None):
         """
-        Add non-evidence related content to the local locker.
+        Add non-evidence text content to the local locker.
 
         :param content: the content to add to the local locker.
         :param folder: the folder in the local locker to add the content to.
         :param filename: the name of the file in the local locker.
         """
-        path = os.path.join(self.local_path, folder)
-        if not os.path.exists(path):
-            os.mkdir(path)
         if not filename:
-            raise ValueError('Filename cannot be empty.')
-        results_file = os.path.join(path, filename)
-        with open(results_file, 'w+') as f:
-            f.write(content)
-        self.repo.index.add([results_file])
+            raise ValueError('You must provide a filename.')
+        path = Path(self.local_path, folder)
+        path.mkdir(parents=True, exist_ok=True)
+        results_file = Path(path, filename)
+        results_file.write_text(content)
+        self.repo.index.add([str(results_file)])
 
     def get_content_from_locker(self, folder='', filename=None):
         """
-        Read non-evidence related content from the local locker.
+        Read non-evidence text content from the local locker.
 
         :param folder: the folder in the local locker to get the content from.
         :param filename: the name of the file in the local locker.
         """
         if not filename:
-            raise ValueError('Filename cannot be empty.')
-        file_path = os.path.join(self.local_path, folder, filename)
+            raise ValueError('You must provide a filename.')
+        file_path = Path(self.local_path, folder, filename)
         content = None
-        if os.path.exists(file_path):
-            with open(file_path) as f:
-                content = f.read()
+        if file_path.is_file():
+            content = file_path.read_text()
         return content
 
     def get_reports_metadata(self):
@@ -680,38 +666,59 @@ class Locker(object):
         :returns: reports metadata.
         """
         metadata = {}
-        rpts = os.walk(os.path.join(self.local_path, 'reports'))
-        for path in [rpath for rpath, _, files in rpts if INDEX_FILE in files]:
-            with open(os.path.join(path, INDEX_FILE), 'r') as f:
-                for rpt_name, rpt_metadata in json.loads(f.read()).items():
-                    rpt_path = os.path.join(
-                        path.split(self.local_path).pop(), rpt_name
-                    )
-                    metadata[rpt_path] = rpt_metadata
+        for idx in Path(self.local_path, 'reports').rglob(INDEX_FILE):
+            for rpt_name, rpt_metadata in json.loads(idx.read_text()).items():
+                rpt_path = idx.relative_to(self.local_path).with_name(rpt_name)
+                metadata[str(rpt_path)] = rpt_metadata
         return metadata
 
     def get_abandoned_evidences(self, threshold=None):
         """
-        Provide a list of evidence where the update ``threshold`` has passed.
+        Provide a set of evidence where the update ``threshold`` has passed.
 
         :param int threshold: the time in seconds after TTL expires that
           evidence can remain un-updated before it is considered abandoned.
           The abandoned evidence threshold defaults to 30 days if none is
           provided.
 
-        :returns: a list of abandoned evidence files.
+        :returns: a set of abandoned evidence file relative paths.
         """
-        abandoned_evidence = []
-        tree = self.repo.head.commit.tree
-        for f in tree.traverse():
-            if (f.type != 'blob' or f.path.startswith('notifications/')
-                    or f.path == 'check_results.json'
-                    or f.path.split('/').pop() in AE_EXEMPT):
-                continue
+        abandoned_evidence = set()
+        for f in self._get_git_files('evidence'):
             metadata = self.get_evidence_metadata(f.path, dt.utcnow())
             if self._evidence_abandoned(metadata, threshold):
-                abandoned_evidence.append(f.path)
-        return set(abandoned_evidence)
+                abandoned_evidence.add(f.path)
+        return abandoned_evidence
+
+    def get_empty_evidences(self):
+        """
+        Provide a list of evidence paths to empty evidence files.
+
+        Evidence content is considered empty based on an evidence object's
+        is_empty property.  This information is stored in evidence metadata.
+
+        :returns: a list of empty evidence file relative paths.
+        """
+        empty_evidence = []
+        for f in self._get_git_files('index'):
+            for ev_name, ev_meta in json.loads(f.data_stream.read()).items():
+                if ev_meta.get('empty', False):
+                    empty_evidence.append(
+                        str(PurePath(f.path).with_name(ev_name))
+                    )
+        return empty_evidence
+
+    def get_large_files(self, size=LF_DEFAULT):
+        """
+        Provide a dictionary of "large" evidence locker files.
+
+        A "large" file is one whose size is > the ``size`` argument provided.
+
+        :param int size: file size threshold.
+
+        :returns: a dictionary of file paths and sizes of "large" files.
+        """
+        return {f.path: f.size for f in self._get_git_files() if f.size > size}
 
     def delete_repo_locally(self):
         """Remove the local git repository."""
@@ -757,12 +764,12 @@ class Locker(object):
 
         :returns: the metadata for the evidence specified.
         """
-        pkg_index_path = self.get_index_file_by_path(evidence_path)
-        if not os.path.exists(pkg_index_path):
+        pkg_index_path = Path(self.get_index_file_by_path(evidence_path))
+        if not pkg_index_path.is_file():
             return
         ev_path, ev_name = evidence_path.rsplit('/', 1)
         if evidence_dt:
-            pkg_index_path = os.path.join(ev_path, INDEX_FILE)
+            pkg_index_path = str(PurePath(ev_path, INDEX_FILE))
             commit = self.get_latest_commit(pkg_index_path, evidence_dt)
             if not commit:
                 return
@@ -770,7 +777,7 @@ class Locker(object):
                 commit.tree[pkg_index_path].data_stream.read()
             )
         else:
-            metadata = json.loads(open(pkg_index_path).read())
+            metadata = json.loads(pkg_index_path.read_text())
         return metadata.get(
             ev_name,
             self._get_partitioned_evidence_metadata(metadata, ev_name)
@@ -800,7 +807,7 @@ class Locker(object):
         return
 
     def _repo_init(self):
-        if os.path.isdir(os.path.join(self.local_path, '.git')):
+        if Path(self.local_path, '.git').is_dir():
             self.logger.info(f'Using locker found in {self.local_path}...')
             self.repo = git.Repo(self.local_path)
         else:
@@ -873,8 +880,8 @@ class Locker(object):
 
     def _get_file_content(self, file_path, file_dt=None, binary_content=False):
         if not file_dt:
-            mode = 'rb' if binary_content else 'r'
-            return open(os.path.join(self.local_path, file_path), mode).read()
+            p = Path(self.local_path, file_path)
+            return p.read_text() if not binary_content else p.read_bytes()
         commit = self.get_latest_commit(file_path, file_dt)
         if not commit:
             raise HistoricalEvidenceNotFoundError(
@@ -894,15 +901,14 @@ class Locker(object):
         if getattr(evidence, 'is_partitioned', False):
             for part_hash in metadata['partitions'].keys():
                 paths.append(
-                    os.path.join(
+                    PurePath(
                         evidence.dir_path, f'{part_hash}_{evidence.name}'
                     )
                 )
         else:
-            paths.append(evidence.path)
+            paths.append(PurePath(evidence.path))
         for path in paths:
-            full_path = os.path.join(self.local_path, path)
-            if not os.path.exists(full_path):
+            if not Path(self.local_path, path).is_file():
                 raise ValueError(
                     f'Evidence {path} was not found in the locker'
                 )
@@ -912,7 +918,52 @@ class Locker(object):
         if not ignore_ttl and ttl_expired:
             raise StaleEvidenceError(f'Evidence {evidence.path} is stale')
 
+    def _get_git_files(self, file_type='all'):
+        iz = {
+            'all': lambda g: g.type == 'blob',
+            'evidence': is_evidence_file,
+            'index': is_index_file
+        }
+        return filter(iz[file_type], self.repo.head.commit.tree.traverse())
 
-def is_index_file(path):
-    """Confirm whether the supplied path is to an index file."""
-    return os.path.basename(path) == INDEX_FILE
+    def _log_large_files(self):
+        large_files = self.get_large_files(
+            get_config().get('locker.large_file_threshold', LF_DEFAULT)
+        )
+        if large_files:
+            msg = ['LARGE FILES (Hosting service may reject due to size):\n']
+            for fpath, size in large_files.items():
+                formatted_size = f'{size/MB:.1f} MB'
+                if formatted_size == '0.0 MB':
+                    formatted_size = f'{str(size)} Bytes'
+                msg.append(f'      {fpath} is {formatted_size}')
+            self.logger.info('\n'.join(msg) + '\n')
+
+
+def is_evidence_file(git_obj):
+    """
+    Confirm whether the supplied git object is an evidence file.
+
+    :param git_obj: A GitPython object
+
+    :returns: True or False (Object is or isn't an evidence file)
+    """
+    return (
+        git_obj.type == 'blob'
+        and not git_obj.path.startswith('notifications/')
+        and git_obj.path != 'check_results.json'
+        and PurePath(git_obj.path).name not in NOT_EVIDENCE
+    )
+
+
+def is_index_file(obj):
+    """
+    Confirm whether the supplied object is, or points to, an index file.
+
+    :param obj: Either a GitPython object or a relative file path as a string
+
+    :returns: True or False (Object is or isn't an index file)
+    """
+    if isinstance(obj, str):
+        return PurePath(obj).name == INDEX_FILE
+    return obj.type == 'blob' and PurePath(obj.path).name == INDEX_FILE